+++ rc_network/src/protocol/packet/entity/head_look.rs
@@ -1,11 +1,11 @@
-use crate::protocol::packet::PacketType;
 use crate::protocol::data::read_types::{read_unsignedbyte, read_varint};
-use std::io::{Cursor};
+use crate::protocol::packet::PacketType;
+use std::io::Cursor;
 
 #[derive(Debug)]
 pub struct EntityHeadLookPacket {
-    pub entity_id: i64,
-    pub yaw: u8
+    pub entity_id: i32,
+    pub yaw: u8,
 }
 
 impl PacketType for EntityHeadLookPacket {
@@ -13,9 +13,6 @@ impl PacketType for EntityHeadLookPacket {
         let entity_id = read_varint(buf);
         let yaw = read_unsignedbyte(buf);
 
-        Box::new(EntityHeadLookPacket {
-            entity_id,
-            yaw
-        })
+        Box::new(EntityHeadLookPacket { entity_id, yaw })
     }
 }
