+++ rc_network/src/protocol/types/ingredient.rs
@@ -1,16 +1,15 @@
+use crate::protocol::data::read_types::read_varint;
 use crate::protocol::types::slot::Slot;
 use std::io::Cursor;
-use crate::protocol::data::read_types::read_varint;
 
 #[derive(Debug)]
 pub struct Ingredient {
-    count: i64,
-    items: Vec<Slot>
+    count: i32,
+    items: Vec<Slot>,
 }
 
 impl Ingredient {
     pub fn deserialize(buf: &mut Cursor<Vec<u8>>) -> Self {
-
         let count = read_varint(buf);
         let mut items = Vec::new();
 
@@ -18,9 +17,6 @@ impl Ingredient {
             items.push(Slot::deserialize(buf));
         }
 
-        Ingredient {
-            count,
-            items
-        }
+        Ingredient { count, items }
     }
 }
